refactor: Entity 클래스 Dto로 변경 및 패키지 변경
