--- src/main/java/chess/db/entity/SquareEntity.java
@@ -1,7 +1,0 @@
-package chess.db.entity;
-
-import chess.domain.piece.Color;
-import chess.domain.piece.PieceType;
-
-public record SquareEntity(String position, PieceType pieceType, Color color) {
-}
