+++ src/main/java/chess/domain/ChessBoard.java
@@ -1,11 +1,11 @@
 package chess.domain;
 
+import chess.dto.SquareDto;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.piece.Position;
 import chess.domain.piece.type.Empty;
-import chess.db.entity.SquareEntity;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.LinkedHashMap;
