+++ src/test/java/chess/dao/SquareDaoTest.java
@@ -9,7 +9,7 @@
 //import chess.domain.piece.Position;
 //import chess.domain.piece.Rank;
 //import chess.domain.piece.type.Bishop;
-//import chess.entity.SquareEntity;
+//import chess.entity.SquareDto;
 //import java.util.List;
 //import java.util.Map;
 //import org.junit.jupiter.api.DisplayName;
@@ -23,12 +23,12 @@
 //    @Test
 //    void findSquares() {
 //        // given && when
-//        final List<SquareEntity> squares = squareDao.findSquares("1");
+//        final List<SquareDto> squares = squareDao.findSquares("1");
 //
 //        // then
 //        assertThat(squares).containsExactlyInAnyOrder(
-//                new SquareEntity("a1", PieceType.ROOK, Color.WHITE),
-//                new SquareEntity("a2", PieceType.KNIGHT, Color.WHITE)
+//                new SquareDto("a1", PieceType.ROOK, Color.WHITE),
+//                new SquareDto("a2", PieceType.KNIGHT, Color.WHITE)
 //        );
 //    }
 //
@@ -38,10 +38,10 @@
 //        // given && when
 //        Map<Position, Piece> pieces = Map.of(new Position(File.A, Rank.THREE), new Bishop(Color.WHITE));
 //        squareDao.addSquares(pieces, "1");
-//        final List<SquareEntity> squares = squareDao.findSquares("1");
+//        final List<SquareDto> squares = squareDao.findSquares("1");
 //
 //        // then
-//        assertThat(squares).contains(new SquareEntity("a3", PieceType.BISHOP, Color.WHITE));
+//        assertThat(squares).contains(new SquareDto("a3", PieceType.BISHOP, Color.WHITE));
 //    }
 //
 //    @DisplayName("해당 게임아이디를 가진 보드칸들이 정상적으로 삭제되는지 확인한다.")
@@ -49,7 +49,7 @@
 //    void deleteSquares() {
 //        // given && when
 //        squareDao.deleteSquares("1");
-//        final List<SquareEntity> squares = squareDao.findSquares("1");
+//        final List<SquareDto> squares = squareDao.findSquares("1");
 //
 //        // then
 //        assertThat(squares).isEmpty();
